LayoutTests/platform/mac-wk1/imported/w3c/web-platform-tests/subresource-integrity/integrity-policy/script.https-expected.txt



FAIL Ensure that a script without integrity did not run assert_unreached: A report of integrity-violation from https://localhost:9443/common/dispatcher/remote-executor.html?pipe=header(Integrity-Policy%2Cblocked-destinations%3D%5C(script%5C)%5C%2C%20endpoints%3D%5C(integrity-endpoint-1%20integrity-endpoint-2%5C))%7Cheader(Integrity-Policy-Report-Only%2Cblocked-destinations%3D%5C(script%5C)%5C%2C%20endpoints%3D%5C(integrity-endpoint-3%5C))%7Cheader(Reporting-Endpoints%2C%20integrity-endpoint-1%3D%22https%3A%2F%2Flocalhost%3A9443%2Freporting%2Fresources%2Freport.py%3FreportID%3D45177130-58cd-4860-8cd4-afb7e70ea02b%22%5C%2C%20integrity-endpoint-2%3D%22https%3A%2F%2Flocalhost%3A9443%2Freporting%2Fresources%2Freport.py%3FreportID%3D511adc8c-04ff-4ef1-b885-f03a343568e9%22%5C%2C%20integrity-endpoint-3%3D%22https%3A%2F%2Flocalhost%3A9443%2Freporting%2Fresources%2Freport.py%3FreportID%3Dd457c930-6fe7-4ae4-ba68-44077e95eb1d%22)&uuid=254b0aa3-7c17-4461-92d7-6264bf4a0455 is not found. Reached unreachable code
FAIL Ensure that a script with unknown integrity algorithm did not run assert_unreached: A report of integrity-violation from https://localhost:9443/common/dispatcher/remote-executor.html?pipe=header(Integrity-Policy%2Cblocked-destinations%3D%5C(script%5C)%5C%2C%20endpoints%3D%5C(integrity-endpoint-1%20integrity-endpoint-2%5C))%7Cheader(Integrity-Policy-Report-Only%2Cblocked-destinations%3D%5C(script%5C)%5C%2C%20endpoints%3D%5C(integrity-endpoint-3%5C))%7Cheader(Reporting-Endpoints%2C%20integrity-endpoint-1%3D%22https%3A%2F%2Flocalhost%3A9443%2Freporting%2Fresources%2Freport.py%3FreportID%3De5b5b272-0ff4-4b28-8fe0-194c726f08df%22%5C%2C%20integrity-endpoint-2%3D%22https%3A%2F%2Flocalhost%3A9443%2Freporting%2Fresources%2Freport.py%3FreportID%3D64a0b2fe-ec42-4d2b-8b03-84d49eae5f7b%22%5C%2C%20integrity-endpoint-3%3D%22https%3A%2F%2Flocalhost%3A9443%2Freporting%2Fresources%2Freport.py%3FreportID%3Dd7665f5b-f03a-43b7-9413-a601587fd0ee%22)&uuid=cedb071c-7957-41b3-94dd-51d6c8018ea3 is not found. Reached unreachable code
FAIL Ensure that a script without integrity algorithm runs and gets reported in report-only mode assert_unreached: A report of integrity-violation from https://localhost:9443/common/dispatcher/remote-executor.html?pipe=%7Cheader(Integrity-Policy-Report-Only%2Cblocked-destinations%3D%5C(script%5C)%5C%2C%20endpoints%3D%5C(integrity-endpoint-3%5C))%7Cheader(Reporting-Endpoints%2C%20integrity-endpoint-1%3D%22https%3A%2F%2Flocalhost%3A9443%2Freporting%2Fresources%2Freport.py%3FreportID%3D7c92d4fe-5b4e-4143-bf88-f1ca3fcee15a%22%5C%2C%20integrity-endpoint-2%3D%22https%3A%2F%2Flocalhost%3A9443%2Freporting%2Fresources%2Freport.py%3FreportID%3D8392380d-c600-4db2-845b-5cc5b2a4a4f9%22%5C%2C%20integrity-endpoint-3%3D%22https%3A%2F%2Flocalhost%3A9443%2Freporting%2Fresources%2Freport.py%3FreportID%3Dfbdca365-e3ae-4297-9107-160528bc944f%22)&uuid=aaf6de10-cfa1-4459-85f5-bce898aaf710 is not found. Reached unreachable code
FAIL Ensure that a no-cors script gets blocked assert_unreached: A report of integrity-violation from https://localhost:9443/common/dispatcher/remote-executor.html?pipe=header(Integrity-Policy%2Cblocked-destinations%3D%5C(script%5C)%5C%2C%20endpoints%3D%5C(integrity-endpoint-1%20integrity-endpoint-2%5C))%7Cheader(Integrity-Policy-Report-Only%2Cblocked-destinations%3D%5C(script%5C)%5C%2C%20endpoints%3D%5C(integrity-endpoint-3%5C))%7Cheader(Reporting-Endpoints%2C%20integrity-endpoint-1%3D%22https%3A%2F%2Flocalhost%3A9443%2Freporting%2Fresources%2Freport.py%3FreportID%3Df79882d0-f999-40c0-8c37-4f47f3606950%22%5C%2C%20integrity-endpoint-2%3D%22https%3A%2F%2Flocalhost%3A9443%2Freporting%2Fresources%2Freport.py%3FreportID%3Dd4a6a1a4-ae50-4ae5-a6e0-b88a497e4850%22%5C%2C%20integrity-endpoint-3%3D%22https%3A%2F%2Flocalhost%3A9443%2Freporting%2Fresources%2Freport.py%3FreportID%3D242123ed-5999-48f6-90bb-90f830663acc%22)&uuid=d24b44b6-7f98-4192-a11e-5c3eb752395c is not found. Reached unreachable code
PASS Ensure that ReportingObserver gets called without endpoints
PASS Ensure that a script with integrity runs
FAIL Ensure that a script with signature integrity runs assert_equals: expected true but got false
PASS Ensure that a data URI script with no integrity runs
PASS Ensure that a no-CORS data URI script with no integrity runs
PASS Ensure that a blob URL script with no integrity runs
PASS Ensure that a no-CORS blob URL script with no integrity runs
